% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD166_19_S1.R
\docType{data}
\name{CD166_19_S1}
\alias{CD166_19_S1}
\title{Itrax core scanner data for core CD166_19_S1}
\format{
A list containing the following:
\describe{
  \item{xrf}{a tibble of scan variables}
  \item{rgb}{a list containing the optical image matrix and associated metadata}
  \item{rad}{a list containing the radiographic image matrix and associated metadata}
  }
}
\source{
Bishop, Thomas; Charidemou, Miros (2023): Core Scanning Data from Core CD166/19. PANGAEA, https://doi.pangaea.de/10.1594/PANGAEA.955347
}
\usage{
CD166_19_S1
}
\description{
A dataset containing XRF, optical and radiographic images and metadata for ocean core CD166_19_S1. This is a subset of a more complete dataset available.
}
\keyword{datasets}
