% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_averaging.r
\name{itrax_averaging}
\alias{itrax_averaging}
\title{Reduce size of Itrax result data}
\usage{
itrax_averaging(dataframe, interval)
}
\arguments{
\item{dataframe}{defines the name of the dataframe to reduce --- this should have been imported using \code{"itrax_import()"}}

\item{interval}{defines the intervals at which to resample, in mm.}
}
\value{
a dataframe of the averaged Itrax data
}
\description{
Resamples Itrax results files into smaller averaged ranges
}
\examples{
itrax_averaging(df, interval = 10)

}
