% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_corescan.r
\name{itrax_corescan}
\alias{itrax_corescan}
\title{Import an Itrax Scan Object}
\usage{
itrax_corescan(image = "WV17-2-D1_rgb_16 bit.tif",
  imagemetadata = "document.txt", radiograph = "radiograph.raw",
  radiographmetadata = "document.txt", xrf = "result.txt",
  xrfmetadata = "document.txt", spectra = "XRF data",
  settings = "settings_240117.dfl")
}
\arguments{
\item{image}{name of the image file}

\item{imagemetadata}{location of the metadata associated with the image scan}

\item{radiograph}{location of the raw radiograph file}

\item{radiographmetadata}{location of the metadata associated with the radiograph scan}

\item{xrf}{location of the xrf results file}

\item{xrfmetadata}{location of the metadata associated with the xrf scan}

\item{spectra}{location of the xrf spectra files}

\item{settings}{location of the settings file associated with the xrf scan}
}
\value{
a object of all the data, where for scan data the row names are positions on the scanner in mm
}
\description{
Brings together all of the data in an Itrax Core Scan in a single object.
}
\examples{
itrax_corescan()

}
