% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_correlation.r
\name{itrax_correlation}
\alias{itrax_correlation}
\title{Calculate a correlation matrix for Itrax result data}
\usage{
itrax_correlation(dataframe, elementsonly = TRUE, zeros = "addone",
  transform = TRUE, diagrams = TRUE)
}
\arguments{
\item{dataframe}{defines the name of the dataframe to reduce --- this should have been imported using \code{"itrax_import()"} or \code{"itrax_join()"}}

\item{elementsonly}{logical operator that if TRUE will only perform the analysis for elemental data}

\item{zeros}{can be either "addone" or "limit" --- this defines what to do with zero values when normalisating. Limit uses 0.001 as the zero value, add one adds one to all data}

\item{transform}{logical operator that if TRUE will center-log-transform the data}

\item{diagrams}{logical operator, if TRUE will produce a graphical output}
}
\value{
a correlation matrix object
}
\description{
Calculates a correlation matrix for Itrax data results including normalisation and visualisation
}
\examples{
itrax_correlation(df)

}
