% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_image.r
\name{itrax_image}
\alias{itrax_image}
\title{Read an Itrax Image File}
\usage{
itrax_image(file = "optical.tif", meta = "document.txt", plot = FALSE)
}
\arguments{
\item{file}{defines the name of the datafile to parse}

\item{meta}{defines the relating metadata}

\item{plot}{would you like to create a plot as a side-effect?}
}
\value{
a matrix of RGB values, and the relevant data from the metadata file relating to the image.
Also computes the aspect ratio of the image.
}
\description{
Reads an Itrax image file and trims it according to the metadata provided.
}
\examples{
\dontrun{itrax_image(file = "optical.tif", meta = "document.txt", plot = TRUE)}

}
