% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_join.R
\name{itrax_join}
\alias{itrax_join}
\title{Join two or more Itrax result datasets}
\usage{
itrax_join(list)
}
\arguments{
\item{list}{a list of dataframes that are parsed Itrax result files --- this should have been imported using \code{"itrax_import()"} and must have a depth variable present
#'}
}
\value{
a dataframe of all the input data
}
\description{
Join two or more Itrax datasets that have been parsed using \code{"itrax_import()"}
}
\examples{
itrax_correlation( list(core1 = df1, core2 = df2, core3 = df3) )

}
