% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_meta.r
\name{itrax_meta}
\alias{itrax_meta}
\title{Parse Itrax scan metadata}
\usage{
itrax_meta(datafile = "document.txt")
}
\arguments{
\item{datafile}{a \code{"document.txt files"} generated from an Itrax core scanner}
}
\value{
a dataframe of all the parsed input data
}
\description{
Parses the \code{"document.txt files"} generated from Itrax core scanners
}
\examples{
itrax_meta(system.file("extdata",
                       "CD166_19_S1_xrf_document.txt",
                       package = "itraxR",
                       mustWork = TRUE))

}
