% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_munsell.r
\name{itrax_munsell}
\alias{itrax_munsell}
\title{Estimate Munsell colour from an Itrax scan}
\usage{
itrax_munsell(filename, scanwidth = 10, munsellator = FALSE,
  intervals = "xrf")
}
\arguments{
\item{scanwidth}{defines the swath width (in mm) to use to generate Munsell colours. This is to remove any core containers or distortion at the edges}

\item{munsellator}{set to TRUE if you are using a Munsellator calibration device. If not, you must supply a colour calibrated image or the output will be nonsense}

\item{intervals}{this is the depth-wise averaging interval. By default it parses the document.txt metadata file and sets this interval equal to the XRF scan interval, but you can supply a figure.}

\item{datafile}{defines the name of the datafile to parse}
}
\value{
a matrix of RGB values and munsell colours
}
\description{
Estimates Munsell colour from a calibrated Itrax scan, or if using a Munsellator block, performs the calibration as well.
}
\examples{
itrax_munsell( filename=system.file( "extdata", "optical.tif", package = "itraxR" ) )

}
