% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_ordination.r
\name{itrax_ordination}
\alias{itrax_ordination}
\title{Principle Component Analysis on Itrax scan data}
\usage{
itrax_ordination(dataframe, elementsonly = TRUE, zeros = "addone",
  transform = TRUE, diagrams = TRUE)
}
\arguments{
\item{dataframe}{pass the name of a dataframe parsed using \code{"itrax_import()"} or \code{"itrax_join()"}}

\item{elementsonly}{binary operator that if TRUE will only perform the analysis for elemental data}

\item{zeros}{can be either "addone" or "limit" --- this defines what to do with zero values when normalisating. Limit uses 0.001 as the zero value, add one adds one to all data}

\item{transform}{binary operator that if TRUE will center-log-transform the data}

\item{diagrams}{binary operator, if TRUE will produce a graphical output, if BIPLOT will only produce a biplot}
}
\value{
an ordination results object
}
\description{
Performs and visualises principle component analysis data from Itrax result data
}
\examples{
itrax_ordination(df)

}
