% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_qspecsettings.R
\name{itrax_qspecsettings}
\alias{itrax_qspecsettings}
\title{Read a Q-Spec settings file and parse the key-value pairs}
\usage{
itrax_qspecsettings(filename = "Results_settings.dfl")
}
\arguments{
\item{filename}{the *.dfl settings file that relates to the rest of the data}
}
\value{
a tibble of the parsed data
}
\description{
This is used to retrieve settings important elsewhere, for example the mca bin width and offset
}
\examples{
itrax_qspecsettings(filename = system.file("extdata",
                                "Results_settings.dfl",
                                package = "itraxR",
                                mustWork = TRUE)
                                )

}
