% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_radiograph_calibrate.r
\name{itrax_radiograph_calibrate}
\alias{itrax_radiograph_calibrate}
\title{Import Itrax core-scanner radiograph file}
\usage{
itrax_radiograph_calibrate(filename, ladder = NULL,
  ladder_position = NULL)
}
\arguments{
\item{datafile}{defines the name of the datafile to parse}
}
\value{
a matrix of the parsed Itrax radiograph
}
\description{
Imports and parses data from a results file created by Q-Spec software, part of the Itrax core scanner.
}
\examples{
itrax_radiograph( filename=system.file( "extdata", "radiograph.tif", package = "itraxR" ) )

}
