% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_restspectra.r
\name{itrax_restspectra}
\alias{itrax_restspectra}
\title{Make a spectrograph from raw Itrax data spectra files}
\usage{
itrax_restspectra(foldername, datapos = 30)
}
\arguments{
\item{foldername}{defines the folder where the spectra \code{"*.spe"} files are located}

\item{datapos}{defines the row at which spectral data begins in the files}
}
\value{
a dataframe of all the spectral data
}
\description{
Parses a folder full of raw spectra files from an Itrax core scanner and produces a spectral graph of all the data by position
}
\examples{
itrax_restspectra( foldername = "XRF Data", datapos = 37 )

}
