% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_section.R
\name{itrax_section}
\alias{itrax_section}
\title{Cluster analysis and statistical grouping of Itrax data}
\usage{
itrax_section(dataframe, divisions = 30, zeros = "addone",
  elements = c(NULL), graph = TRUE)
}
\arguments{
\item{dataframe}{defines Itrax results data parsed using \code{"itrax_import()"} or \code{"itrax_join()"}}

\item{divisions}{defines the number of sub-groups to divide the data into. Use this to define the number of samples for calibration.}

\item{zeros}{can be either "addone" or "limit" --- this defines what to do with zero values when normalisating. Limit uses 0.001 as the zero value, add one adds one to all data}

\item{elements}{an optional list of elements to include, otheriwse, all elements present are used}

\item{graph}{a binary operator that if TRUE will produce a graphical representation of the results}
}
\value{
an object that includes the depth (or position) and the respective group, alongside a list of group center samples for choosing calibration samples
}
\description{
Performs a cluster analysis and automatic statistical grouping of parsed Itrax results data
}
\examples{
itrax_section( df )

}
