% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_spectracompare.R
\name{itrax_spectracompare}
\alias{itrax_spectracompare}
\title{Import Itrax core-scanner spectra file comparison}
\usage{
itrax_spectracompare(filea, fileb, datapos = 37, graph = TRUE)
}
\arguments{
\item{filea}{defines the name of the first file}

\item{fileb}{defines the name of the second file}

\item{datapos}{defines the line of the input files at which the metadata ends and the spectral data begins.}

\item{graph}{binary operator that if TRUE will produce a graph of the two spectra}
}
\value{
returns a dataframe of the difference between the two
}
\description{
Imports, parses and compares two Itrax spectra files, used for Q.C. on Itrax calibration blocks.
}
\examples{
itrax_spectracompare(filea = system.file("extdata", "a.spe", package = "itraxR"), fileb = system.file("extdata", "b.spe", package = "itraxR"))

}
