% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_import.R
\name{multivariate_import}
\alias{multivariate_import}
\title{Prepare data for multivariate analysis}
\usage{
multivariate_import(dataframe, elementsonly, zeros, transform)
}
\arguments{
\item{dataframe}{pass the name of a dataframe parsed using \code{"itrax_import()"} or \code{"itrax_join()"}}

\item{elementsonly}{if TRUE, only chemical elements are included. If FALSE, the data is passed unfiltered, otherwise a character vector of desired variable names can be supplied}

\item{zeros}{if "addone", adds one to all values. If "limit", replaces zero values with 0.001. Otherwise a function can be supplied to remove zero values.}

\item{transform}{binary operator that if TRUE will center-log-transform the data, if FALSE will leave the data untransformed. Otherwise, a function can be supplied to transform the data.}
}
\description{
does all the preparation work for multivariate methods
}
\keyword{internal}
