% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid_labeller.R
\name{uid_labeller}
\alias{uid_labeller}
\title{Generate uids for datasets}
\usage{
uid_labeller(dataframe)
}
\arguments{
\item{dataframe}{pass the name of a dataframe parsed using \code{"itrax_import()"} or \code{"itrax_join()"}}
}
\description{
Assigns unique identifiers to dataset observations. 
If there is already a column called uid, that does infact contain unique ids, it simply uses them. 
Otherwise, it looks for the columns `depth` and `label`, and generates a meaningful identifier by combining them.
Otherwise, it simply uses consecutive numbers prefixed by `uid`.
}
\keyword{internal}
