% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD166_19_S1.R
\docType{data}
\name{CD166_19_S1}
\alias{CD166_19_S1}
\title{Itrax core scanner data for core CD166_19_S1}
\format{
A list containing the following:
\describe{
  \item{xrf}{a tibble of scan variables}
  \item{rgb}{a list containing the optical image matrix and associated metadata}
  \item{rad}{a list containing the radiographic image matrix and associated metadata}
  }

  @source Wynn, R. B., and B. T. Cronin. 2005. “RRS "Charles Darwin" Cruise CD166, 29 Oct - 22 Nov 2004. Sedimentary processes and deposits in the Agadir Basin and Gulf of Cadiz.” 59. Vol. 44.
}
\usage{
CD166_19_S1
}
\description{
A dataset containing XRF, optical and radiographic images and metadata for ocean core CD166_19_S1.
}
\keyword{datasets}
