% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_image.r
\name{itrax_image}
\alias{itrax_image}
\title{Estimate Munsell colour from an Itrax scan}
\usage{
itrax_image(filename, metadata = "document.txt")
}
\arguments{
\item{filename}{defines the name of the datafile to parse}

\item{metadata}{defines the relating metadata}
}
\value{
a matrix of RGB values
}
\description{
Estimates Munsell colour from a calibrated Itrax scan, or if using a Munsellator block, performs the calibration as well.
}
\examples{
itrax_image()

}
