% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_image.r
\name{itrax_image}
\alias{itrax_image}
\title{Import Itrax core-scanner RGB image file}
\usage{
itrax_image(filename)
}
\arguments{
\item{filename}{defines the name of the datafile to parse}
}
\value{
a matrix of the parsed Itrax image
}
\description{
Imports and parses an image file from an Itrax core-scanner.
}
\examples{
itrax_image( filename=system.file( "extdata", "optical.tif", package = "itraxR" ) )

}
