% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_join.r
\name{itrax_join}
\alias{itrax_join}
\title{Join two or more Itrax result datasets}
\usage{
itrax_join(list)
}
\arguments{
\item{list}{a list of dataframes that are parsed Itrax result files --- this should have been imported using \code{"itrax_import()"} and must have a depth variable present. This should take the form \code{"list(core1 = core1, core2 = core2)"}}
}
\value{
a tibble of all the input data
}
\description{
Join two or more Itrax datasets that have been parsed using \code{"itrax_import()"}
}
\examples{
itrax_join(list(core1 = CD166_19_S1$xrf, core2 = CD166_19_S1$xrf))
}
