% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_munsell.r
\name{itrax_munsell}
\alias{itrax_munsell}
\title{Estimate Munsell colour from an Itrax scan}
\usage{
itrax_munsell(filename, scanwidth = 300, munsellator = FALSE)
}
\arguments{
\item{datafile}{defines the name of the datafile to parse}
}
\value{
a matrix of RGB values and munsell colours
}
\description{
Estimates Munsell colour from a calibrated Itrax scan, or if using a Munsellator block, performs the calibration as well.
}
\examples{
itrax_munsell( filename=system.file( "extdata", "optical.tif", package = "itraxR" ) )

}
