% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrax_radiograph.r
\name{itrax_radiograph}
\alias{itrax_radiograph}
\title{Read an Itrax Radiograph File}
\usage{
itrax_radiograph(file = "radiograph.tif", meta = "document.txt", plot = FALSE)
}
\arguments{
\item{plot}{would you like to create a plot as a side-effect?}

\item{filename}{defines the name of the datafile to parse}

\item{metadata}{defines the relating metadata}
}
\value{
a matrix of RGB values, and the relevant data from the metadata file relating to the image.
Also computes the aspect ratio of the image.
}
\description{
Reads an Itrax radiograph file and trims it according to the metadata provided.
}
\examples{
\dontrun{itrax_radiograph(file = "radiograph.tif", meta = "document.txt", plot = TRUE)}

}
